refactor: Move networking classes to 'api' folder
