+++ remote-config/src/test/kotlin/com/rakuten/tech/mobile/remoteconfig/api/HeadersInterceptorSpec.kt
@@ -1,5 +1,6 @@
-package com.rakuten.tech.mobile.remoteconfig
+package com.rakuten.tech.mobile.remoteconfig.api
 
+import com.rakuten.tech.mobile.remoteconfig.RobolectricBaseSpec
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.mockwebserver.MockResponse
